client/requirements.txt
certifi==2019.11.28
grpcio==1.27.2
grpcio-tools==1.27.2
protobuf==3.11.3
pygame==1.9.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
six==1.14.0
wincertstore==0.2
